% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cashflowPost}
\alias{cashflowPost}
\title{Example data for a livelong life risk insurance with payoff 100
in case of death}
\format{
A data.table with columns time, from, to, amount:
\describe{
\item{time} {the time step the cashflow applies to}
\item{from} {the state at time step t}
\item{to} {the state at time step t + 1}
\item{amount} {the amount of cashflow}
}
}
\usage{
cashflowPost
}
\description{
Example data for a livelong life risk insurance with payoff 100
in case of death
}
\keyword{datasets}
