% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cashflowPre}
\alias{cashflowPre}
\title{Example data for a livelong life risk insurance contract
with premium 10 to be paid when the policy holder is alive}
\format{
A data.table with columns time, state, amount:
\describe{
\item{time} {the time step the cashflow applies to}
\item{state} {the state for which the cashflow arises}
\item{amount} {the amount of cashflow}
}
}
\usage{
cashflowPre
}
\description{
Example data for a livelong life risk insurance contract
with premium 10 to be paid when the policy holder is alive
}
\keyword{datasets}
