% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeDist.R
\name{completeDist}
\alias{completeDist}
\title{A function to compute the distrubtion of the present value of
future cashflows}
\usage{
completeDist(mc, granularity = 0.1)
}
\arguments{
\item{granularity}{indicates the amount of precision to be used in the distribution}
}
\description{
A function to compute the distrubtion of the present value of
future cashflows
}
