% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeV.R
\name{completeV}
\alias{completeV}
\title{The function completeV creates the expected value of future cashflows
for all possible states and times.}
\usage{
completeV(mc)
}
\description{
The function completeV creates the expected value of future cashflows
for all possible states and times.
}
