% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardDist.R
\name{forwardDist}
\alias{forwardDist}
\title{evaluate the distribution function for one given state, timee and value}
\usage{
forwardDist(u, state, time, trans, payoffPost, payoffPre, i)
}
\arguments{
\item{u}{value for which to evalute the distribution function}

\item{state}{state for which to evalute the distribution function}

\item{time}{time for which to evalute the distribution function}

\item{trans}{a tidy data.table containg the transition probabilities}

\item{payoffPost}{a tidy data.table containing the payoff of kind post}

\item{payoffPre}{a tidy data.table containing the payoff of kind pre
export}
}
\description{
This function constructs a tree of points that are to be stepwise
evaluated for higher time-steps.
}
