% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans}
\alias{trans}
\title{Example data from Statistik Oesterreich}
\format{
A data.table with columns time, from, to, p
\describe{
\item{time} {the time step in question}
\item{from} {the state at time t}
\item{to} {the state at time t+1}
\item{p}{the probability for the change to happen}
}
}
\source{
table Ausfuehrliche, allgemeine und ausgeglichene Sterbetafeln 1868/71 bis 2010/12 nach dem Geschlecht
 \url{https://www.statistik.at/web_de/statistiken/menschen_und_gesellschaft/bevoelkerung/sterbetafeln/index.html}
}
\usage{
trans
}
\description{
The data.table contains yearly mortality data
}
\keyword{datasets}
