% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardDist.R
\name{forwardDist}
\alias{forwardDist}
\title{evaluate the distribution function for one given state, timee and value}
\usage{
forwardDist(mc, u, state, time)
}
\arguments{
\item{u}{value for which to evalute the distribution function}

\item{state}{state for which to evalute the distribution function}

\item{time}{time for which to evalute the distribution function}
}
\description{
This function constructs a tree of points that are to be stepwise
evaluated for higher time-steps.
}
